requirements.txt
streamlit
geopandas
shapely
fiona
pyproj
langchain
langchain-groq
openai
python-dotenv
